feat(hud): セッション終了後も対戦相手の表示を保持する
離席・観戦・フィルター変更・セッション終了で対戦相手の集計HUDとドリルダウンを消さず、直近ハンドを終了直後に振り返れる状態を維持する。

信頼できる新着席lineupとテーブル切替では従来どおり置換し、現在ハンド専用のSPR・ポットオッズだけを終了時にローカルでクリアする。
