+++ README.md
@@ -14,7 +14,7 @@ PokerChase HUDは、ゲーム中に受信したイベントから各プレイヤ
 - **プレイヤータイプ判定**: 対戦相手を統計に基づいて🦈 TAG、💣 LAG、🪨 ニット、🐟 フィッシュ、🐳 ホエールに分類します。
 - **プレイヤー別の詳細表示**: ポジション別統計と、最近のハンドをHUDから確認できます。直近ハンドは表示件数を10・25・50・100件から選べ、「参加のみ」（既定ON）でブラインドを取られただけのハンドを隠せます（どちらの選択も端末ごとに保存されます）。そのプレイヤーが実際に配られたハンドだけを、ホールカード（ランクのみ表示・スートは4色で判別、正確な表記はマウスオーバー。自分のハンドはショーダウンに関わらず常に表示、相手はショーダウンで公開されたもの）・プリフロップのライン（`OR2.2`=2.2BBへオープンレイズ、`3B9`=9BBへ3bet、`CC2.2`/`3CC9`=オープン/3betへのコールドコールのように、一般的なHUDの短縮表記でBB単位の額つき）・ボード・フロップ以降のアクション（ベット／レイズはポットに対する比率つき。`XC/B33` のように表示し、実額はマウスオーバーで確認できます）・BB単位の損益とあわせて一覧します。
 - **ゲーム開始前の自分の統計**: プレイヤー情報が保存済みなら、最初のハンドが配られる前から自分の通算統計を表示します。
-- **離席プレイヤーの表示保持**: バストやテーブル移動で席を離れたプレイヤーの統計を、席が交代するかセッションが終わるまで薄く表示します。
+- **離席プレイヤーの表示保持**: バストなどで席を離れたプレイヤーの統計を席が交代するまで薄く表示し、セッション終了後も統計や直近ハンドを振り返れます。
 - **リアルタイム計算**: 全プレイヤーのポットオッズとSPR、自分のハンド改善確率をアクションごとに更新します。
 - **ハンド履歴**: PokerStars形式のハンドログをリアルタイムに表示・出力します。表示位置や大きさも調整できます。
 - **フィルター**: ゲーム種別、テーブル人数、直近ハンド数で集計対象を切り替えられます。
