+++ src/background/event-ingestion.ts
@@ -9,7 +9,7 @@ import PokerChaseService, {
   isApplicationApiEvent
 } from '../app'
 import { autoSyncService } from '../services/auto-sync-service'
-import { connectedPorts, startPortPing, setLastKnownStats } from './ports'
+import { clearLatestRealTimeStats, connectedPorts, startPortPing } from './ports'
 import {
   INVALID_API_TYPE_ID_BUCKET,
   recordUndecodedEvent
@@ -439,6 +439,12 @@ const processEvent = async (
   // 詳細は`applySessionActivity`のコメント参照。
   applySessionActivity(rawApiTypeId, message, false, port)
 
+  if (rawApiTypeId === ApiType.EVT_SESSION_RESULTS) {
+    // 集計lineupはレビュー用に保持する一方、現在ハンド専用の値は終了済み。
+    // パース不能な309でも後続のフィルター再計算へ混入しないようrawで消す。
+    clearLatestRealTimeStats()
+  }
+
   // リプレイ取り込み層（既定OFF）。**セッション中は取得しない**（MUST）ので、
   // ここでやるのは HandId をキューへ積むことだけ。取得はセッション終了の
   // トリガーから走る（`replay-import.ts`）。
@@ -462,62 +468,6 @@ const processEvent = async (
       (message as { HandId?: unknown }).HandId,
       Date.now()
     ).catch(err => console.error('[background] Replay import enqueue failed:', err))
-  }
-
-  if (rawApiTypeId === ApiType.EVT_SESSION_RESULTS) {
-    // #179 round3指摘: セッション終了(EVT_SESSION_RESULTS)によるHUDクリアは
-    // App.tsx側のReact stateだけで完結しており、background(ports.ts)の
-    // `lastKnownStats`はセッションをまたいで残り続ける。この状態で
-    // Popupのバトルタイプフィルターが変更されると、message-router.tsの
-    // `updateBattleTypeFilter`ハンドラーが`getLastKnownStats()`（終了済み
-    // lineupのまま）を使って`service.statsOutputStream.write(...)`を
-    // 再トリガーし、ブロードキャストで終了済みlineupが復活してApp.tsxの
-    // クリア済みパネルへ再度流し込まれてしまう。上と同じ「パース成功後の
-    // data.ApiTypeIdではなく生メッセージの数値ApiTypeIdだけを見る」
-    // raw-firstパターンで（309のペイロード破壊的変更に影響されないよう）
-    // ここでlastKnownStatsを空にしておけば、以降のフィルター変更は
-    // `lastKnownStats.length > 0`のガードに引っかからずセッション開始前と
-    // 同じ「何もブロードキャストしない」挙動になる。プリゲーム・ヒーロー
-    // スタッツの復元（#158, `requestLatestStats`→`getLatestSessionStats`）
-    // はDBを読む別経路でありlastKnownStatsを参照しないため、この変更の
-    // 影響を受けない。
-    //
-    // post-merge reviewでは一時ここにhero単独lineupを合成する修正
-    // （round4/round5）や、App.tsx側にセッション状態・ヒーロー身元の
-    // 検証機構を足す修正（round6の「相互作用マトリクス」設計）を
-    // 積んだが、いずれもオーナー判断で撤回されている（PR #191,
-    // 2026-07-20, sola「それほど重要な機能ではないので、bで十分です」）。
-    // 理由: `service.setBattleTypeFilter()`は内部で無条件に
-    // `ReadEntityStream.recalculateStats()`を呼び、`lastKnownStats`の
-    // 中身に関係なく`service.latestEvtDeal.SeatUserIds`（ヒーローの
-    // 直近の実在席時点のフルの顔ぶれ。セッション終了後もクリアされ
-    // ない）を再計算・再ブロードキャストしてしまう。つまりこのファイル
-    // 側で`lastKnownStats`をどう作っても、その再ブロードキャストは
-    // 止められない別経路であり、追いかけるだけ複雑化する一方だった。
-    //
-    // 採用した方針（保守的な縮小スコープ）: bust後のミュート表示・
-    // hero身元の保持は「連続したライブシーケンス内」でのみ保証する
-    // （#158のセッション終了時hero保持は例外的にApp.tsx側で維持）。
-    // セッション終了後にフィルターを変更すると、`recalculateStats()`
-    // がヒーロー在籍時点の最後の実テーブル（対戦相手を含む）を新
-    // フィルターで再表示することがあるが、これは「不正確なデータ」
-    // ではなく「文脈的に古い可能性のある正確なデータ」であり、この
-    // 機能の重要度に見合わないため許容する。ここは元のround3の
-    // 意図通り単純な`[]`のままにする -- `updateBattleTypeFilter`の
-    // 明示的な`getLastKnownStats()`ベースの再write()を単にno-opに
-    // するだけで、他には何もしない。
-    //
-    // なお#188（read-entity-stream.tsのrecalculateStats()を呼ぶ
-    // message-router.tsのupdateBattleTypeFilterハンドラー自身）で、
-    // このwrite()と`recalculateStats()`の競合（観戦中のlineupが
-    // ヒーロー在籍dealのevtDealとペアリングされてしまうケース）に
-    // 対する`lineup-identity`ガードが別途入っており、この単純な
-    // `[]`のままでも競合は起きない。
-    //
-    // 203(参加取消申込)はここに含めない: 303/308が一度も届いていない
-    // （ハンドが一度も始まっていない）ため、そもそもクリアすべき
-    // ライブlineupが存在しない。
-    setLastKnownStats([])
   }
 
   // Auto-sync起動・保留中アップデートの安全性再チェックも、上のセッション状態
