+++ src/content_script.ts
@@ -353,17 +353,16 @@ window.addEventListener('message', (event: MessageEvent<unknown>) => {
         isGameActive = false
         stopKeepalive()
       }
-      // App.tsx へセッション終了を通知（bustしたプレイヤーの薄暗い表示を含む、
-      // hero以外の全HUDパネルをクリアするため）。309はここで生イベントとして
+      // App.tsx へセッション終了を通知し、現在ハンド専用のリアルタイム統計を
+      // クリアする。集計HUDはレビュー用に保持する。309はここで生イベントとして
       // 既に観測済みなので、background往復の新チャネルを追加せずその場でdispatchする。
       window.dispatchEvent(new CustomEvent(POKER_CHASE_SESSION_END_EVENT))
       break
 
     case EVT_ENTRY_CANCELLED_API_TYPE_ID:
       // 参加取消: ハンドが一度も始まっていないので、309と違いApp.tsxへの
-      // セッション終了通知（POKER_CHASE_SESSION_END_EVENT）は不要
-      // （そもそもクリアすべきライブHUDが存在しない）。keepaliveの解除
-      // だけ行う。
+      // セッション終了通知（POKER_CHASE_SESSION_END_EVENT）は不要。
+      // keepaliveの解除だけ行う。
       if (isGameActive) {
         isGameActive = false
         stopKeepalive()
@@ -426,13 +425,6 @@ const requestLatestStats = (preGame = false) => {
 }
 
 const messageHandlers: Record<string, (message: ChromeMessage) => void> = {
-  updateBattleTypeFilter: (message) => {
-    if ('filterOptions' in message) {
-      window.dispatchEvent(new CustomEvent(EVENTS.UPDATE_BATTLE_TYPE_FILTER, {
-        detail: message.filterOptions
-      }))
-    }
-  },
   latestStats: (message) => {
     if ('stats' in message) {
       const data: StatsData = { stats: message.stats }
